Laravel notify done by suash

+++ app/Http/Controllers/HelpCenterController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use App\Models\HelpCenter;
 use Illuminate\Http\Request;
 
@@ -9,7 +10,7 @@ class HelpCenterController extends Controller
     protected $helpCenter = null;
     public function __construct(HelpCenter $helpCenter)
     {
-       $this->helpCenter = $helpCenter;
+        $this->helpCenter = $helpCenter;
     }
     /**
      * Display a listing of the resource.
@@ -18,8 +19,8 @@ class HelpCenterController extends Controller
      */
     public function index()
     {
-        $helpCenters = HelpCenter::orderBy('order_id' , 'ASC')->get();
-        return view('admin.helpCenter.helpCenterList')->with('helpCenter_data',$helpCenters);
+        $helpCenters = HelpCenter::orderBy('order_id', 'ASC')->get();
+        return view('admin.helpCenter.helpCenterList')->with('helpCenter_data', $helpCenters);
     }
 
     /**
@@ -48,6 +49,11 @@ class HelpCenterController extends Controller
         $data['order_id'] = getOrderId($order_id);
         $this->helpCenter->fill($data);
         $status = $this->helpCenter->save();
+        if ($status) {
+            notify()->success('Help added successfully.');
+        } else {
+            notify()->error('Sorry! There was problem while adding help.');
+        }
         return redirect()->route('helpCenter.index');
     }
 
@@ -59,8 +65,9 @@ class HelpCenterController extends Controller
      */
     public function show($id)
     {
-        $this->helpCenter= $this->helpCenter->find($id);
+        $this->helpCenter = $this->helpCenter->find($id);
         if (!$this->helpCenter) {
+            notify()->error('This help doesnot exists');
             return redirect()->route('helpCenter.index');
         }
         return view('admin.helpCenter.helpCenterView')->with('helpCenter_data', $this->helpCenter);
@@ -74,8 +81,9 @@ class HelpCenterController extends Controller
      */
     public function edit($id)
     {
-    $this->helpCenter = $this->helpCenter->find($id);
-     if (!$this->helpCenter) {
+        $this->helpCenter = $this->helpCenter->find($id);
+        if (!$this->helpCenter) {
+            notify()->error('This help doesnot exists');
             return redirect()->route('helpCenter.index');
         }
         return view('admin.helpCenter.helpCenterForm')->with('helpCenter_data', $this->helpCenter);
@@ -90,16 +98,21 @@ class HelpCenterController extends Controller
      */
     public function update(Request $request, $id)
     {
-    $this->helpCenter= $this->helpCenter->find($id);
-    if (!$this->helpCenter) {
-
-        return redirect()->route('helpCenter.index');
+        $this->helpCenter = $this->helpCenter->find($id);
+        if (!$this->helpCenter) {
+            notify()->error('This help doesnot exists');
+            return redirect()->route('helpCenter.index');
         }
         $rules = $this->helpCenter->getRules();
         $request->validate($rules);
         $data = $request->all();
         $this->helpCenter->fill($data);
         $status = $this->helpCenter->save();
+        if($status){
+            notify()->success('Help updated successfully.');
+        }else{
+            notify()->error('Sorry! There was problem while adding help.');
+        }
         return redirect()->route('helpCenter.index');
     }
 
@@ -111,8 +124,19 @@ class HelpCenterController extends Controller
      */
     public function destroy($id)
     {
-    $this->helpCenter = $this->helpCenter->find($id);
-    $del = $this->helpCenter->delete();
-    return redirect()->route('helpCenter.index');
+        $this->helpCenter = $this->helpCenter->find($id);
+        if (!$this->helpCenter) {
+            notify()->error('This help doesnot exists');
+            redirect()->route('helpCenter.index');
+        }
+        $del = $this->helpCenter->delete();
+        if ($del) {
+            notify()->success('Help deleted successfully');
+        }
+        else {
+            //message
+            notify()->error('Sorry! there was problem in deleting help');
+        }
+        return redirect()->route('helpCenter.index');
     }
 }
